requirements.txt
Django==4.0
djangorestframework==3.12.4
channels==3.0.4
pytz==2021.3
